+++ settings.py
@@ -45,7 +45,7 @@ APP_KITS = {
             'task1_tr2',
             'task1_results_tr12',
             'task2_instructions',
-            'new_task2_practice_chp',
+            'task2_prac_tr2',
             'new_task2_quiz',
             'task2_tr2',
             'new_task2_results',
@@ -102,7 +102,7 @@ SESSION_CONFIGS = [
                       'tr3_dynamic_pooled_results',
                       'task2_instructions',
                       'task2_prac_tr13',
-                      'new_task2_practice_chp',
+                      'task2_prac_tr2',
                       'new_task2_quiz',
                       'task2_tr1',
                       'task2_tr2',
