+++ task2_prac_tr2/models.py
@@ -14,7 +14,7 @@ Your app description
 
 
 class Constants(BaseConstants):
-    name_in_url = 'new_task2_practice_chp'
+    name_in_url = 'task2_prac_tr2'
     players_per_group = None
     num_rounds = 1
     initial_wealth = 100
