+++ task2_prac_tr2/templates/task2_prac_tr2/.DS_Store
